+++ wayland-server/src/lib.rs
@@ -124,9 +124,9 @@ pub mod sources {
     // event loop, other than the wayland protocol sockets.
 
     pub use event_sources::{FdEventSource, FdEventSourceImpl, FdInterest};
+    pub use event_sources::{IdleEventSource, IdleEventSourceImpl};
     pub use event_sources::{SignalEventSource, SignalEventSourceImpl};
     pub use event_sources::{TimerEventSource, TimerEventSourceImpl};
-    pub use event_sources::{IdleEventSource, IdleEventSourceImpl};
 }
 
 /// Common routines for wayland resource objects.